Refactor the redux store to use immer library
The key idea here is that it forces you to edit the draft object, which is not the immutable state.  It's a pretty low overhead system that makes it hard to do it wrong.  I also cleaned up a few more ANYs in code andshwowed how to use the redux concept of actions, types, and reducers.